hdoc working properly
