+++ Makefile
@@ -6,7 +6,7 @@
 #    By: zmakhkha <zmakhkha@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/03/10 14:52:21 by zmakhkha          #+#    #+#              #
-#    Updated: 2023/06/19 16:57:34 by zmakhkha         ###   ########.fr        #
+#    Updated: 2023/06/21 11:11:21 by zmakhkha         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -26,6 +26,8 @@ HEADERS = header.h
 #parsing part:
 SRC_MN =	main.c \
 			common_files/ast/ft_ast_utils.c \
+			common_files/ast/ft_asthdoc.c \
+			common_files/ast/ft_asthdoc2.c \
 			common_files/ast/ft_leafs_utils.c \
 			common_files/ast/ft_main_ast.c \
 			common_files/ast/ft_ops_utils.c \
